Topology: deduplicate show joining minix and controller.
