+++ app/controllers/subnet_topology_controller.rb
@@ -1,6 +1,4 @@
 class SubnetTopologyController < TopologyController
-  include TopologyMixin
-
   before_action :check_privileges
   before_action :get_session_data
   after_action :cleanup_action
